Implement genetic algorithm
This is quite a cumbersome push, I know. I was not entirely sure how to go
about developing the algorithm at first. Bits and pieces were constructed
here and there. Things were moved around. Nothing was really presentable or
complete until it all fit together. At the end of the day, it works, and
the code doesn't look too bad.